feat:  added validations for category

+++ src/controllers/CategoryController.js
@@ -12,11 +12,23 @@ module.exports = {
 
     const category = await Category.findByPk(id);
 
+    if(!category) {
+      return res.status(404).json({ error: 'Category not found!'})
+    }
+
     return res.json(category);
   },
 
   async store(req, res) {
     const { name, description } = req.body;
+
+    const categoryAlreadyExists = await Category.findOne({
+      where: { name }
+    });
+
+    if(categoryAlreadyExists) {
+      return res.status(400).json({ error: 'Category already exists!'})
+    }
 
     const category = await Category.create({ name, description });
 
@@ -26,8 +38,20 @@ module.exports = {
   async delete(req, res) {
     const { id } = req.params;
 
-    await Category.destroy({ where: { id } });
+    const category = await Category.findByPk(id);
 
+    if(!category) {
+      return res.status(400).json({ error: 'Category not found!'})
+    }
+
+    try {
+      await Category.destroy({ where: { id } });
+    } catch (error) {
+      return res.status(400).json({ 
+        error: 'Category cannot be removed as it contains movies!'
+      })
+    }
+    
     return res.send();
   }
 }
